Python/E11/requirements.txt
requests==2.25.1
lxml==4.6.3
beautifulsoup4==4.10.0
Pillow==8.3.2
